Documentation for the project
Added a pdf file explaining the entire breakdown and workflow of the Quantum anomaly detection project and findings and conclusion of the project. The file discusses the current works in the form of Literature Review and discusses the module-wise breakdown and workflow of data along with the future scope and used references.